requirements.txt
spotipy>=2.23.0
yt-dlp>=2024.10.22
rich>=13.7.0
python-dotenv>=1.0.1
